requirements.txt
annotated-types==0.7.0
blinker==1.9.0
click==8.2.1
Flask==2.3.3
flasgger==0.9.7.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.0
pytest==8.4.1
python-dotenv==1.0.0
pytz==2025.2
SQLAlchemy==2.0.23
typing-inspection==0.4.1
typing_extensions==4.14.1
Werkzeug==3.1.3
